Replace blocking-read with drop-frames
blocking-read has been already renamed to drop-frames but it's
functionality was still present.

The blocking-read functionality was completely removed. Setting it to
false resulted in the same frame being integrated multiple times. Which
is not what we want in a dense SLAM system. The new behavior is as if
blocking-read is true, integrating each frame at most once.

The drop-frames parameter will be used in a future commit to allow
skipping frames when the input framerate is higher than the processing
framerate. It will be set to not drop frames by default.