Rename Node::data() to Node::setData() where appropriate
It makes it easier for both the programmer and the compiler to
differentiate between the data() that gets and the data() that sets.
This is because VoxelData data(const Eigen::Vector3i&, const int);
can match void data(const Eigen::Vector3i&, const VoxelData&); if
VoxelData is a POD (int, float, etc.).