Merge benchmark into main
This reduces the number of executables from 2*4*3=24 to 2*4*2=16.

The benchmark functionality of showing stats is now available when
passing --no-gui. The previous function of showing just the frame and
camera position has been removed as this information is present in the
new output.

As a temporary measure until we refactor the python benchmark scripts,
the multi-column output of --no-gui is converted to the 4-column output
that the scripts expect.