Add static_assert for near and far planes
Ensures that nearPlane < farPlane at compile time. No need to test this
in the raycasting implementations so the ifs were removed. Git can't
figure out the minimal diff but the only changes made were removing the
ifs and de-indenting the code they contained.

Once we make it possible to set the near and far planes at runtime, we
should move this test (and probably convert it to a runtime check that
errors out with a message) to the code that reads their values.