Refactor allocation and integration
`near_plane` and `far_plane` check as well as  block's `active` update via the `is_visible` bool have been moved from the `projective_functor` to the updating function of each voxel implementation. This allows more flexibility for the use of the `projective_functor` and makes it independent of the type of input image used (e.g. depth or rgba).